requirements.txt
appdirs==1.4.2
configparser==3.5.0
ecdsa==0.13
future==0.16.0
packaging==16.8
pycrypto==2.6.1
PyGithub==1.32
pyparsing==2.1.10
python-jose==1.3.2
six==1.10.0
